% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcee_helper_utils.R
\name{.mcee_assert_df}
\alias{.mcee_assert_df}
\title{Assert that input is a data frame}
\usage{
.mcee_assert_df(data)
}
\arguments{
\item{data}{Object to check}
}
\value{
Invisibly TRUE if valid, otherwise stops with error
}
\description{
Assert that input is a data frame
}
