% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcee_helper_utils.R
\name{.mcee_build_f_matrix}
\alias{.mcee_build_f_matrix}
\title{Build basis matrix f(t) from time-varying effect formula}
\usage{
.mcee_build_f_matrix(time_varying_effect_form, data)
}
\arguments{
\item{time_varying_effect_form}{RHS-only formula}

\item{data}{Data frame to evaluate formula on}
}
\value{
Model matrix with basis functions evaluated at each row
}
\description{
Build basis matrix f(t) from time-varying effect formula
}
