% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcee_helper_nuisance_config.R
\name{.mcee_check_formula_mediator}
\alias{.mcee_check_formula_mediator}
\title{Check config formula for inclusion/exclusion of mediator}
\usage{
.mcee_check_formula_mediator(config, target, mediator)
}
\arguments{
\item{config}{A nuisance config list (may or may not contain `formula`).}

\item{target}{Character scalar, one of "p", "q", "eta", "mu", "nu".}

\item{mediator}{Character scalar: mediator variable name.}
}
\value{
Invisibly TRUE. Warnings are produced if formulas look suspicious.
}
\description{
Check config formula for inclusion/exclusion of mediator
}
