% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcee_helper_utils.R
\name{.mcee_check_id_rows_grouped}
\alias{.mcee_check_id_rows_grouped}
\title{Check that rows for each subject appear in contiguous blocks}
\usage{
.mcee_check_id_rows_grouped(data, id, max_show = 5)
}
\arguments{
\item{data}{Data frame}

\item{id}{Column name for subject ID}

\item{max_show}{Maximum number of offending IDs to show in error message}
}
\value{
Invisibly TRUE if valid, otherwise stops with error
}
\description{
Check that rows for each subject appear in contiguous blocks
}
