% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mcee.R
\name{.mcee_compact_model_info}
\alias{.mcee_compact_model_info}
\title{Generate compact one-line description of nuisance model object}
\usage{
.mcee_compact_model_info(obj)
}
\arguments{
\item{obj}{Fitted model object or known value descriptor}
}
\value{
Character string describing the object
}
\description{
Generate compact one-line description of nuisance model object
}
