% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcee_helper_nuisance_config.R
\name{.mcee_default_family}
\alias{.mcee_default_family}
\title{Select default GLM family based on nuisance parameter type}
\usage{
.mcee_default_family(target, method)
}
\arguments{
\item{target}{Nuisance parameter name ("p", "q", "eta", "mu", "nu")}

\item{method}{Learning method name}
}
\value{
GLM family object or NULL for non-GLM methods
}
\description{
Select default GLM family based on nuisance parameter type
}
