% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcee_helper_utils.R
\name{.mcee_drop_var_from_rhs}
\alias{.mcee_drop_var_from_rhs}
\title{Remove a variable from RHS-only formula}
\usage{
.mcee_drop_var_from_rhs(rhs_only_formula, var)
}
\arguments{
\item{rhs_only_formula}{RHS-only formula}

\item{var}{Variable name to remove}
}
\value{
Modified formula with variable removed
}
\description{
Remove a variable from RHS-only formula
}
