% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcee_helper_utils.R
\name{.mcee_message_if_no_availability_provided}
\alias{.mcee_message_if_no_availability_provided}
\title{Print informative message if no availability column provided}
\usage{
.mcee_message_if_no_availability_provided(availability, verbose)
}
\arguments{
\item{availability}{Availability column name (or NULL)}

\item{verbose}{Logical; whether to print messages}
}
\value{
Invisibly TRUE
}
\description{
Print informative message if no availability column provided
}
