% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcee_helper_nuisance_config.R
\name{.mcee_validate_method}
\alias{.mcee_validate_method}
\title{Validate that learning method is supported}
\usage{
.mcee_validate_method(method)
}
\arguments{
\item{method}{Method name to validate}
}
\value{
Invisibly TRUE if valid, otherwise stops with error
}
\description{
Validate that learning method is supported
}
