% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcls.R
\name{wcls}
\alias{wcls}
\title{Estimates the causal excursion effect for continuous outcome MRT}
\usage{
wcls(
  data,
  id,
  outcome,
  treatment,
  rand_prob,
  moderator_formula,
  control_formula,
  availability = NULL,
  numerator_prob = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data set in long format.}

\item{id}{The subject id variable.}

\item{outcome}{The outcome variable.}

\item{treatment}{The binary treatment assignment variable.}

\item{rand_prob}{The randomization probability variable.}

\item{moderator_formula}{A formula for the moderator variables. This should
start with ~ followed by the moderator variables. When set to \code{~ 1}, a fully
marginal excursion effect (no moderators) is estimated.}

\item{control_formula}{A formula for the control variables. This should
start with ~ followed by the control variables. When set to \code{~ 1}, only an
intercept is included as the control variable.}

\item{availability}{The availability variable. Use the default value (\code{NULL})
if your MRT doesn't have availability considerations.}

\item{numerator_prob}{Either a number between 0 and 1, or a variable name for
a column in data. If you are not sure what this is, use the default value (\code{NULL}).}

\item{verbose}{If default (`TRUE`), additional messages will be printed
during data preprocessing.}
}
\value{
An object of type "wcls_fit"
}
\description{
Returns the estimated causal excursion effect (on additive scale) and the estimated standard error.
Small sample correction using the "Hat" matrix in the variance estimate is implemented.
}
\examples{
wcls(
    data = data_mimicHeartSteps,
    id = "userid",
    outcome = "logstep_30min",
    treatment = "intervention",
    rand_prob = 0.6,
    moderator_formula = ~1,
    control_formula = ~logstep_pre30min,
    availability = "avail",
    numerator_prob = 0.6
)
}
